requirements.txt
google-genai
google-adk
pydantic
python-dotenv
litellm